Created a Contact base class so that Address, Email, Phone can extend from it.
